\name{leek_plots}
\alias{leek.pch}
\alias{leek.colors}
\title{
  Essential defaults for making plots in LEEK group publications
}
\description{
  If you write papers with the LEEK group you better use these
  functions, or ... QUACK!
}
\details{
There is only one useful value of \code{pch} in R: \code{leek.pch}.

There is only one good color palette in R and it is given by this
vector: \code{leek.colors}. You know you need to use it or ... QUACK!!
}
\examples{
leek.pch
leek.colors
}
